Fixed: creating entry file failed if parent had been removed.
This happened for example if the app's cache directory had been cleared.

A simpler solution would've been to just add the parent check before all
writes, but ended with a check-and-possibly-mkdir-on-failure approach
instead. Couldn't figure out how to catch the error when using streams
(`File.openWrite()`) so switched to using `RandomAccessFile`s.